+++ src/test/converter/inherit-param-doc/inherit-param-doc.ts
@@ -4,11 +4,18 @@ export interface Base {
      * @param b - Parameter B.
      */
     method1(a: number, b: string): void;
+    /**
+     * @param a - Parameter A.
+     * @param b - Parameter B.
+     */
+    method2(a: number, b: string): void;
 }
 
 export class Class1 implements Base {
     /** @inheritDoc */
     method1(a: number, b: string): void {}
+    /** @inheritDoc */
+    method2(): void {}
 }
 
 export class Class2 implements Base {
@@ -18,6 +25,12 @@ export class Class2 implements Base {
      * @param a - Custom parameter A doc.
      */
     method1(a: number, b: string): void {}
+    /**
+     * @inheritDoc
+     *
+     * @param a - Custom parameter A doc.
+     */
+    method2(a: number): void {}
 }
 
 export class Class3 implements Base {
@@ -27,4 +40,5 @@ export class Class3 implements Base {
      * @param c - Custom second parameter doc with name change.
      */
     method1(a: number, c: string): void {}
+    method2(a: number): void {}
 }